[Tweak] Add co2 targets option (support intensity and budget) and extent (currently only all).
Targets are assumed to be defined in a .csv file with cols "case,type,extent,target" where case is scenario name (e.g. high_high), type is intensity/budget, extent is all and target is the co2 target in kt (kilo tonnes).

Update submodule to GAMS code that supports the above.

Add code needed for co2 targets to be modelled as cumulative across all zones (all) or per zone (zonal)